pyproject.toml
[tool.poetry]
name = "hotel-project"
version = "0.1.0"
description = ""
authors = ["tetianaklius <tetyanaklyus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
mysqlclient = "^2.2.7"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
celery = "^5.5.2"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
pillow = "^11.1.0"
redis = "^6.0.0"
better-profanity = "^0.7.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
